requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
motor==3.5.1
pymongo[srv]==4.8.0
pydantic==2.8.2
pydantic-settings==2.4.0
python-multipart==0.0.9
httpx==0.27.2
reportlab==4.2.2
python-dotenv==1.0.1
boto3==1.35.0
aiofiles==24.1.0
